+++ Clause/Having.php
@@ -288,7 +288,7 @@ trait Having
      */
     public function havingContains(Closure|StatementInterface|string $column, string $value): static
     {
-        return $this->andHaving($column, 'LIKE', sprintf('%%%s%%', $value));
+        return $this->andHaving($column, 'LIKE', sprintf('%%%s%%', $this->escapeLike($value)));
     }
 
     /**
@@ -301,7 +301,7 @@ trait Having
      */
     public function havingStartsWith(Closure|StatementInterface|string $column, string $value): static
     {
-        return $this->andHaving($column, 'LIKE', sprintf('%s%%', $value));
+        return $this->andHaving($column, 'LIKE', sprintf('%s%%', $this->escapeLike($value)));
     }
 
     /**
@@ -314,6 +314,18 @@ trait Having
      */
     public function havingEndsWith(Closure|StatementInterface|string $column, string $value): static
     {
-        return $this->andHaving($column, 'LIKE', sprintf('%%%s', $value));
+        return $this->andHaving($column, 'LIKE', sprintf('%%%s', $this->escapeLike($value)));
+    }
+
+    /**
+     * Escape LIKE wildcard characters.
+     *
+     * @param string $value
+     *
+     * @return string
+     */
+    private function escapeLike(string $value): string
+    {
+        return str_replace(['\\', '%', '_'], ['\\\\', '\\%', '\\_'], $value);
     }
 }